requirements.txt
flask
mss
pyautogui
pyperclip
pillow
